@@ -21,27 +21,26 @@
     "@leaselab/shared-utils": "*",
     "@leaselab/storage-cloudflare": "*",
     "@leaselab/storage-core": "*",
-    "@remix-run/cloudflare": "^2.15.0",
-    "@remix-run/cloudflare-pages": "^2.15.0",
-    "@remix-run/react": "^2.15.0",
-    "isbot": "^5.1.17",
-    "openai": "^4.70.0",
+    "@remix-run/cloudflare": "^2.17.2",
+    "@remix-run/cloudflare-pages": "^2.17.2",
+    "@remix-run/react": "^2.17.2",
+    "isbot": "^5.1.32",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-dropzone": "^14.3.8",
-    "zod": "^3.22.4"
+    "zod": "^4.1.13"
   },
   "devDependencies": {
-    "@cloudflare/workers-types": "^4.20241106.0",
-    "@remix-run/dev": "^2.15.0",
+    "@cloudflare/workers-types": "^4.20251128.0",
+    "@remix-run/dev": "^2.17.2",
+    "@tailwindcss/postcss": "^4.1.17",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
-    "autoprefixer": "^10.4.20",
-    "postcss": "^8.4.47",
-    "tailwindcss": "^3.4.13",
-    "typescript": "^5.3.3",
-    "vite": "^5.4.11",
-    "vite-tsconfig-paths": "^5.1.3",
-    "wrangler": "^3.87.0"
+    "autoprefixer": "^10.4.22",
+    "postcss": "^8.5.6",
+    "tailwindcss": "^4.1.17",
+    "typescript": "^5.9.3",
+    "vite": "^5.4.21",
+    "vite-tsconfig-paths": "^5.1.4",
+    "wrangler": "^4.51.0"
   }
 }
