@@ -4,7 +4,8 @@
   "private": true,
   "type": "module",
   "workspaces": [
-    "apps/*"
+    "apps/*",
+    "shared/*"
   ],
   "scripts": {
     "dev:site": "npm run dev --workspace=@leaselab/site",
@@ -20,13 +21,19 @@
     "test:coverage": "vitest run --coverage"
   },
   "devDependencies": {
-    "@edge-runtime/vm": "^4.0.4",
-    "@vitest/coverage-v8": "^4.0.10",
+    "@edge-runtime/vm": "^5.0.0",
+    "@vitest/coverage-v8": "^1.6.1",
     "npm-run-all": "^4.1.5",
     "typescript": "^5.3.3",
-    "vitest": "^4.0.10"
+    "vitest": "^1.6.1"
   },
   "engines": {
-    "node": ">=18.0.0"
+    "node": ">=20.0.0"
+  },
+  "overrides": {
+    "esbuild": "^0.27.0",
+    "valibot": "^1.2.0",
+    "estree-util-value-to-estree": "^1.3.0",
+    "vite-node": "^3.2.4"
   }
 }
