@@ -2,11 +2,11 @@ export type { DatabaseInput } from './helpers';
 export { normalizeDb } from './helpers';
 export { getProperties, getPropertyById, getPropertyBySlug, getPropertyWithUnits, createProperty, updateProperty, deleteProperty, getPublicListings, } from './properties';
 export { getUnits, getUnitsByPropertyId, getUnitById, getUnitWithDetails, createUnit, updateUnit, deleteUnit, getUnitHistory, createUnitHistory, } from './units';
-export { getImagesByEntity, getImageById, createImage, updateImage, deleteImage, setCoverImage, } from './images';
-export { getLeads, getLeadById, createLead, updateLead, getLeadFiles, createLeadFile, getAIEvaluation, createAIEvaluation, getLeadHistory, recordLeadHistory, } from './leads';
+export { getImagesByEntity, getImageById, createImage, updateImage, deleteImage, setCoverImage, addImageUrls, getImagesByEntityWithUrls, addImageUrlsWithVerification, getImagesByEntityWithVerification, } from './images';
+export { getLeads, getLeadById, createLead, updateLead, getLeadFiles, createLeadFile, createTempLeadFile, associateFilesWithLead, countLeadFiles, getAIEvaluation, createAIEvaluation, getLeadHistory, recordLeadHistory, } from './leads';
 export { getTenants, getTenantById, } from './tenants';
 export { getWorkOrders, getWorkOrderById, createWorkOrder, updateWorkOrder, deleteWorkOrder, } from './work-orders';
-export { getUserByEmail, getUserById, getUsers, updateUserPassword, updateUserProfile, getUserSiteAccess, getUserAccessibleSites, userHasAccessToSite, grantSiteAccess, revokeSiteAccess, isUserSuperAdmin, setSuperAdminStatus, } from './users';
+export { getUserByEmail, getUserById, getUsers, createUser, updateUserPassword, updateUserProfile, updateUserRole, getUserSiteAccess, getUserAccessibleSites, userHasAccessToSite, grantSiteAccess, revokeSiteAccess, isUserSuperAdmin, setSuperAdminStatus, } from './users';
 export type { AccessibleSite } from './users';
 export { getSiteApiTokens, getSiteApiTokenById, createSiteApiToken, updateSiteApiToken, deleteSiteApiToken, } from './site-tokens';
 export type { SiteApiToken } from './site-tokens';
