@@ -1,4 +1,4 @@
-import { generateId } from '../../../../shared/utils';
+import { generateId } from '~/shared/utils';
 import { normalizeDb } from './helpers';
 // Mapper functions
 function mapLeadFromDb(row) {
@@ -25,8 +25,11 @@ function mapLeadFromDb(row) {
 function mapLeadWithOccupancyFromDb(row) {
     const lead = mapLeadFromDb(row);
     const r = row;
-    lead.isUnitOccupied = Boolean(r.is_unit_occupied);
-    return lead;
+    return {
+        ...lead,
+        isUnitOccupied: Boolean(r.is_unit_occupied),
+        propertyName: r.property_name,
+    };
 }
 function mapLeadFileFromDb(row) {
     const r = row;
@@ -59,14 +62,16 @@ function mapAIEvaluationFromDb(row) {
 export async function getLeads(dbInput, siteId, options) {
     const db = normalizeDb(dbInput);
     const { status, propertyId, sortBy = 'created_at', sortOrder = 'desc', limit = 50, offset = 0 } = options || {};
-    // Join with units to check if the unit is occupied
+    // Join with units to check if the unit is occupied and properties to get property name
     let query = `
     SELECT
       l.*,
       u.status as unit_status,
-      CASE WHEN u.status = 'occupied' THEN 1 ELSE 0 END as is_unit_occupied
+      CASE WHEN u.status = 'occupied' THEN 1 ELSE 0 END as is_unit_occupied,
+      p.name as property_name
     FROM leads l
     LEFT JOIN units u ON l.unit_id = u.id OR (l.property_id = u.property_id AND u.unit_number = 'Main')
+    LEFT JOIN properties p ON l.property_id = p.id
     WHERE l.site_id = ?
   `;
     const params = [siteId];
@@ -78,7 +83,17 @@ export async function getLeads(dbInput, siteId, options) {
         query += ' AND l.property_id = ?';
         params.push(propertyId);
     }
-    const orderColumn = sortBy === 'aiScore' ? 'l.ai_score' : `l.${sortBy.replace(/([A-Z])/g, '_$1').toLowerCase()}`;
+    // Map sort fields to database columns
+    let orderColumn;
+    if (sortBy === 'aiScore' || sortBy === 'ai_score') {
+        orderColumn = 'l.ai_score';
+    }
+    else if (sortBy === 'propertyName' || sortBy === 'property_name') {
+        orderColumn = 'p.name';
+    }
+    else {
+        orderColumn = `l.${sortBy.replace(/([A-Z])/g, '_$1').toLowerCase()}`;
+    }
     query += ` ORDER BY ${orderColumn} ${sortOrder.toUpperCase()} LIMIT ? OFFSET ?`;
     params.push(limit, offset);
     const results = await db.query(query, params);
@@ -179,6 +194,60 @@ export async function createLeadFile(dbInput, siteId, data) {
     VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)
   `, [id, siteId, data.leadId, data.fileType, data.fileName, data.fileSize, data.mimeType, data.r2Key, now]);
     return { ...data, id, uploadedAt: now };
+}
+/**
+ * Create a temporary lead file (before lead is associated)
+ * Used during the upload workflow where files are uploaded before lead submission
+ * Files are staged in staged_files table until associated with a lead
+ */
+export async function createTempLeadFile(dbInput, siteId, data) {
+    const db = normalizeDb(dbInput);
+    const id = generateId('file');
+    const now = new Date().toISOString();
+    // Insert into staging table (will be moved to lead_files when lead is created)
+    await db.execute(`
+    INSERT INTO staged_files (id, site_id, file_type, file_name, file_size, mime_type, r2_key, uploaded_at)
+    VALUES (?, ?, ?, ?, ?, ?, ?, ?)
+  `, [id, siteId, data.fileType, data.fileName, data.fileSize, data.mimeType, data.r2Key, now]);
+    return { id, uploadedAt: now };
+}
+/**
+ * Associate temporary files with a lead
+ * Moves files from staged_files to lead_files and returns the count of files associated
+ */
+export async function associateFilesWithLead(dbInput, siteId, leadId, fileIds) {
+    if (fileIds.length === 0)
+        return 0;
+    const db = normalizeDb(dbInput);
+    const placeholders = fileIds.map(() => '?').join(',');
+    // Move files from staged_files to lead_files
+    // First, insert into lead_files
+    await db.execute(`
+    INSERT INTO lead_files (id, site_id, lead_id, file_type, file_name, file_size, mime_type, r2_key, uploaded_at)
+    SELECT id, site_id, ?, file_type, file_name, file_size, mime_type, r2_key, uploaded_at
+    FROM staged_files
+    WHERE id IN (${placeholders})
+      AND site_id = ?
+  `, [leadId, ...fileIds, siteId]);
+    // Then, delete from staged_files
+    const result = await db.execute(`
+    DELETE FROM staged_files
+    WHERE id IN (${placeholders})
+      AND site_id = ?
+  `, [...fileIds, siteId]);
+    return (result.meta?.rows_written || 0);
+}
+/**
+ * Count files for a lead (or count staged files if leadId is null)
+ */
+export async function countLeadFiles(dbInput, siteId, leadId = null) {
+    const db = normalizeDb(dbInput);
+    const query = leadId === null
+        ? 'SELECT COUNT(*) as count FROM staged_files WHERE site_id = ?'
+        : 'SELECT COUNT(*) as count FROM lead_files WHERE site_id = ? AND lead_id = ?';
+    const params = leadId === null ? [siteId] : [siteId, leadId];
+    const result = await db.queryOne(query, params);
+    return result?.count || 0;
 }
 // AI Evaluations
 export async function getAIEvaluation(dbInput, siteId, leadId) {
