@@ -1,5 +1,5 @@
-import type { Lead, LeadFile, LeadAIResult, LeadHistory } from '../../../../shared/types';
-import { generateId } from '../../../../shared/utils';
+import type { Lead, LeadFile, LeadAIResult, LeadHistory } from '~/shared/types';
+import { generateId } from '~/shared/utils';
 import type { DatabaseInput } from './helpers';
 import { normalizeDb } from './helpers';
 
@@ -26,11 +26,14 @@ function mapLeadFromDb(row: unknown): Lead {
     };
 }
 
-function mapLeadWithOccupancyFromDb(row: unknown): Lead {
+function mapLeadWithOccupancyFromDb(row: unknown): Lead & { isUnitOccupied?: boolean; propertyName?: string } {
     const lead = mapLeadFromDb(row);
     const r = row as Record<string, unknown>;
-    lead.isUnitOccupied = Boolean(r.is_unit_occupied);
-    return lead;
+    return {
+        ...lead,
+        isUnitOccupied: Boolean(r.is_unit_occupied),
+        propertyName: r.property_name as string | undefined,
+    };
 }
 
 function mapLeadFileFromDb(row: unknown): LeadFile {
@@ -74,14 +77,16 @@ export async function getLeads(dbInput: DatabaseInput, siteId: string, options?:
     const db = normalizeDb(dbInput);
     const { status, propertyId, sortBy = 'created_at', sortOrder = 'desc', limit = 50, offset = 0 } = options || {};
 
-    // Join with units to check if the unit is occupied
+    // Join with units to check if the unit is occupied and properties to get property name
     let query = `
     SELECT
       l.*,
       u.status as unit_status,
-      CASE WHEN u.status = 'occupied' THEN 1 ELSE 0 END as is_unit_occupied
+      CASE WHEN u.status = 'occupied' THEN 1 ELSE 0 END as is_unit_occupied,
+      p.name as property_name
     FROM leads l
     LEFT JOIN units u ON l.unit_id = u.id OR (l.property_id = u.property_id AND u.unit_number = 'Main')
+    LEFT JOIN properties p ON l.property_id = p.id
     WHERE l.site_id = ?
   `;
     const params: (string | number)[] = [siteId];
@@ -96,7 +101,15 @@ export async function getLeads(dbInput: DatabaseInput, siteId: string, options?:
         params.push(propertyId);
     }
 
-    const orderColumn = sortBy === 'aiScore' ? 'l.ai_score' : `l.${sortBy.replace(/([A-Z])/g, '_$1').toLowerCase()}`;
+    // Map sort fields to database columns
+    let orderColumn: string;
+    if (sortBy === 'aiScore' || sortBy === 'ai_score') {
+        orderColumn = 'l.ai_score';
+    } else if (sortBy === 'propertyName' || sortBy === 'property_name') {
+        orderColumn = 'p.name';
+    } else {
+        orderColumn = `l.${sortBy.replace(/([A-Z])/g, '_$1').toLowerCase()}`;
+    }
     query += ` ORDER BY ${orderColumn} ${sortOrder.toUpperCase()} LIMIT ? OFFSET ?`;
     params.push(limit, offset);
 
@@ -212,6 +225,75 @@ export async function createLeadFile(dbInput: DatabaseInput, siteId: string, dat
   `, [id, siteId, data.leadId, data.fileType, data.fileName, data.fileSize, data.mimeType, data.r2Key, now]);
 
     return { ...data, id, uploadedAt: now };
+}
+
+/**
+ * Create a temporary lead file (before lead is associated)
+ * Used during the upload workflow where files are uploaded before lead submission
+ * Files are staged in staged_files table until associated with a lead
+ */
+export async function createTempLeadFile(dbInput: DatabaseInput, siteId: string, data: {
+    fileType: LeadFile['fileType'];
+    fileName: string;
+    fileSize: number;
+    mimeType: string;
+    r2Key: string;
+}): Promise<{ id: string; uploadedAt: string }> {
+    const db = normalizeDb(dbInput);
+    const id = generateId('file');
+    const now = new Date().toISOString();
+
+    // Insert into staging table (will be moved to lead_files when lead is created)
+    await db.execute(`
+    INSERT INTO staged_files (id, site_id, file_type, file_name, file_size, mime_type, r2_key, uploaded_at)
+    VALUES (?, ?, ?, ?, ?, ?, ?, ?)
+  `, [id, siteId, data.fileType, data.fileName, data.fileSize, data.mimeType, data.r2Key, now]);
+
+    return { id, uploadedAt: now };
+}
+
+/**
+ * Associate temporary files with a lead
+ * Moves files from staged_files to lead_files and returns the count of files associated
+ */
+export async function associateFilesWithLead(dbInput: DatabaseInput, siteId: string, leadId: string, fileIds: string[]): Promise<number> {
+    if (fileIds.length === 0) return 0;
+
+    const db = normalizeDb(dbInput);
+    const placeholders = fileIds.map(() => '?').join(',');
+
+    // Move files from staged_files to lead_files
+    // First, insert into lead_files
+    await db.execute(`
+    INSERT INTO lead_files (id, site_id, lead_id, file_type, file_name, file_size, mime_type, r2_key, uploaded_at)
+    SELECT id, site_id, ?, file_type, file_name, file_size, mime_type, r2_key, uploaded_at
+    FROM staged_files
+    WHERE id IN (${placeholders})
+      AND site_id = ?
+  `, [leadId, ...fileIds, siteId]);
+
+    // Then, delete from staged_files
+    const result = await db.execute(`
+    DELETE FROM staged_files
+    WHERE id IN (${placeholders})
+      AND site_id = ?
+  `, [...fileIds, siteId]);
+
+    return (result.meta?.rows_written || 0) as number;
+}
+
+/**
+ * Count files for a lead (or count staged files if leadId is null)
+ */
+export async function countLeadFiles(dbInput: DatabaseInput, siteId: string, leadId: string | null = null): Promise<number> {
+    const db = normalizeDb(dbInput);
+    const query = leadId === null
+        ? 'SELECT COUNT(*) as count FROM staged_files WHERE site_id = ?'
+        : 'SELECT COUNT(*) as count FROM lead_files WHERE site_id = ? AND lead_id = ?';
+    const params = leadId === null ? [siteId] : [siteId, leadId];
+
+    const result = await db.queryOne(query, params);
+    return (result as any)?.count || 0;
 }
 
 // AI Evaluations
