@@ -6,7 +6,9 @@
  */
 import { Hono } from 'hono';
 import { authMiddleware } from '../middleware/auth';
-import { getPropertyById, createLead, getPublicListings, getUnitWithDetails, } from '../lib/db';
+import { getPropertyById, createLead, getPublicListings, getUnitWithDetails, createTempLeadFile, countLeadFiles, getImagesByEntityWithVerification, } from '../lib/db';
+import { FILE_UPLOAD_CONSTRAINTS } from '../../../shared/config';
+import { generateId } from '../../../shared/utils';
 const publicRoutes = new Hono();
 // Apply auth middleware to all public routes
 publicRoutes.use('*', authMiddleware);
@@ -20,13 +22,6 @@ publicRoutes.use('*', authMiddleware);
  */
 publicRoutes.get('/properties', async (c) => {
     try {
-        if (!c.env.R2_PUBLIC_URL) {
-            console.error('R2_PUBLIC_URL is not defined');
-            return c.json({
-                error: 'Configuration error',
-                message: 'R2_PUBLIC_URL is not configured',
-            }, 500);
-        }
         const siteId = c.get('siteId');
         const city = c.req.query('city');
         const status = c.req.query('status');
@@ -35,7 +30,7 @@ publicRoutes.get('/properties', async (c) => {
             filters.city = city;
         if (status)
             filters.status = status;
-        const listings = await getPublicListings(c.env.DB, siteId, filters, c.env.R2_PUBLIC_URL);
+        const listings = await getPublicListings(c.env.DB, siteId, filters, c.env.R2_PUBLIC_URL, c.env.PUBLIC_BUCKET);
         return c.json({
             success: true,
             data: listings,
@@ -51,7 +46,7 @@ publicRoutes.get('/properties', async (c) => {
 });
 /**
  * GET /api/public/properties/:id
- * Fetch a single property by ID or slug
+ * Fetch a single property by ID or slug with images (including URLs)
  */
 publicRoutes.get('/properties/:id', async (c) => {
     try {
@@ -64,9 +59,14 @@ publicRoutes.get('/properties/:id', async (c) => {
                 message: 'Property not found',
             }, 404);
         }
+        // Get property images with URLs and verify they exist in R2
+        const images = await getImagesByEntityWithVerification(c.env.DB, siteId, 'property', property.id, c.env.R2_PUBLIC_URL, c.env.PUBLIC_BUCKET);
         return c.json({
             success: true,
-            data: property,
+            data: {
+                ...property,
+                images,
+            },
         });
     }
     catch (error) {
@@ -80,6 +80,7 @@ publicRoutes.get('/properties/:id', async (c) => {
 /**
  * GET /api/public/units/:id
  * Fetch a single unit by ID including property details and images
+ * Returns combined gallery: property images first, then unit images
  */
 publicRoutes.get('/units/:id', async (c) => {
     try {
@@ -92,9 +93,28 @@ publicRoutes.get('/units/:id', async (c) => {
                 message: 'Unit not found',
             }, 404);
         }
+        // Get property images first (if property exists), then unit images
+        // Verify all images exist in R2 before serving
+        let propertyImages = [];
+        if (unit.property) {
+            propertyImages = await getImagesByEntityWithVerification(c.env.DB, siteId, 'property', unit.property.id, c.env.R2_PUBLIC_URL, c.env.PUBLIC_BUCKET);
+            // Mark property images for identification
+            propertyImages = propertyImages.map(img => ({ ...img, isPropertyImage: true }));
+        }
+        // Get unit images with verification
+        const unitImages = await getImagesByEntityWithVerification(c.env.DB, siteId, 'unit', id, c.env.R2_PUBLIC_URL, c.env.PUBLIC_BUCKET);
+        // Combine images: property first, then unit (as per PRD requirements)
+        const allImages = [...propertyImages, ...unitImages];
         return c.json({
             success: true,
-            data: unit,
+            data: {
+                ...unit,
+                images: allImages,
+                property: unit.property ? {
+                    ...unit.property,
+                    images: propertyImages,
+                } : undefined,
+            },
         });
     }
     catch (error) {
@@ -148,6 +168,121 @@ publicRoutes.get('/site-config', async (c) => {
     }
 });
 /**
+ * POST /api/public/leads/files/upload
+ * Upload a file for a lead application (before lead is created)
+ *
+ * Multi-layer validation:
+ * 1. Content-Length header check
+ * 2. MIME type validation
+ * 3. Stream size validation during upload
+ * 4. Post-upload verification
+ */
+publicRoutes.post('/leads/files/upload', async (c) => {
+    try {
+        const siteId = c.get('siteId');
+        // Layer 1: Content-Length header validation
+        const contentLength = c.req.header('content-length');
+        if (!contentLength || parseInt(contentLength) > FILE_UPLOAD_CONSTRAINTS.maxFileSize) {
+            return c.json({
+                error: 'File too large',
+                message: `File size exceeds maximum of ${FILE_UPLOAD_CONSTRAINTS.maxFileSize / 1024 / 1024}MB`,
+                maxSize: FILE_UPLOAD_CONSTRAINTS.maxFileSize,
+            }, 413);
+        }
+        // Parse multipart form data
+        const formData = await c.req.formData();
+        const file = formData.get('file');
+        if (!file || typeof file === 'string') {
+            return c.json({ error: 'Invalid file' }, 400);
+        }
+        const fileType = formData.get('fileType');
+        if (!file) {
+            return c.json({
+                error: 'Validation error',
+                message: 'No file provided',
+            }, 400);
+        }
+        if (!fileType) {
+            return c.json({
+                error: 'Validation error',
+                message: 'fileType is required',
+            }, 400);
+        }
+        // Layer 2: MIME type validation
+        if (!FILE_UPLOAD_CONSTRAINTS.allowedMimeTypes.includes(file.type)) {
+            return c.json({
+                error: 'Unsupported file type',
+                message: `File type ${file.type} is not allowed`,
+                allowedTypes: FILE_UPLOAD_CONSTRAINTS.allowedMimeTypes,
+            }, 415);
+        }
+        // Layer 3: File size validation (redundant check on File object)
+        if (file.size > FILE_UPLOAD_CONSTRAINTS.maxFileSize) {
+            return c.json({
+                error: 'File too large',
+                message: `File size exceeds maximum of ${FILE_UPLOAD_CONSTRAINTS.maxFileSize / 1024 / 1024}MB`,
+                maxSize: FILE_UPLOAD_CONSTRAINTS.maxFileSize,
+            }, 413);
+        }
+        // Check file count limit (count temp files for this site)
+        const currentFileCount = await countLeadFiles(c.env.DB, siteId, null);
+        if (currentFileCount >= FILE_UPLOAD_CONSTRAINTS.maxFilesPerLead * 10) { // Allow some buffer for multiple applicants
+            return c.json({
+                error: 'Too many pending files',
+                message: 'Too many temporary files. Please submit your application or wait.',
+            }, 429);
+        }
+        // Generate R2 key for staged file
+        const fileId = generateId('file');
+        const r2Key = `${siteId}/_staged/${fileId}-${file.name}`;
+        // Upload to R2 PRIVATE_BUCKET
+        await c.env.PRIVATE_BUCKET.put(r2Key, file.stream(), {
+            httpMetadata: {
+                contentType: file.type,
+            },
+        });
+        // Layer 4: Post-upload verification
+        const uploadedObject = await c.env.PRIVATE_BUCKET.head(r2Key);
+        if (!uploadedObject) {
+            throw new Error('File upload failed - file not found after upload');
+        }
+        if (uploadedObject.size > FILE_UPLOAD_CONSTRAINTS.maxFileSize) {
+            // Delete oversized file
+            await c.env.PRIVATE_BUCKET.delete(r2Key);
+            return c.json({
+                error: 'File too large',
+                message: 'File size exceeds maximum after upload',
+            }, 413);
+        }
+        // Create temp file record in database
+        const tempFile = await createTempLeadFile(c.env.DB, siteId, {
+            fileType: fileType,
+            fileName: file.name,
+            fileSize: uploadedObject.size,
+            mimeType: file.type,
+            r2Key,
+        });
+        const response = {
+            fileId: tempFile.id,
+            fileName: file.name,
+            fileSize: uploadedObject.size,
+            fileType: fileType,
+            uploadedAt: tempFile.uploadedAt,
+        };
+        return c.json({
+            success: true,
+            data: response,
+        }, 201);
+    }
+    catch (error) {
+        console.error('Error uploading file:', error);
+        return c.json({
+            error: 'Internal server error',
+            message: error instanceof Error ? error.message : 'Unknown error',
+        }, 500);
+    }
+});
+/**
  * POST /api/public/leads
  * Submit a tenant application
  */
@@ -165,7 +300,48 @@ publicRoutes.post('/leads', async (c) => {
                 }, 400);
             }
         }
+        // Create the lead
         const lead = await createLead(c.env.DB, siteId, body);
+        // Associate uploaded files with the lead (if any)
+        if (body.fileIds && Array.isArray(body.fileIds) && body.fileIds.length > 0) {
+            // First, get the staged files to get their R2 keys
+            const placeholders = body.fileIds.map(() => '?').join(',');
+            const stagedFilesResult = await c.env.DB.prepare(`SELECT id, file_name, r2_key FROM staged_files WHERE id IN (${placeholders}) AND site_id = ?`).bind(...body.fileIds, siteId).all();
+            const stagedFiles = stagedFilesResult.results || [];
+            // Move files in R2 from _staged/ to leads/{leadId}/ and build new key mapping
+            const fileKeyMapping = {};
+            for (const stagedFile of stagedFiles) {
+                const oldKey = stagedFile.r2_key;
+                const newKey = `${siteId}/leads/${lead.id}/${stagedFile.id}-${stagedFile.file_name}`;
+                fileKeyMapping[stagedFile.id] = newKey;
+                // Copy file to new location
+                const object = await c.env.PRIVATE_BUCKET.get(oldKey);
+                if (object) {
+                    await c.env.PRIVATE_BUCKET.put(newKey, object.body, {
+                        httpMetadata: object.httpMetadata,
+                    });
+                    // Delete old staged file
+                    await c.env.PRIVATE_BUCKET.delete(oldKey);
+                }
+            }
+            // Move files in database from staged_files to lead_files with updated R2 keys
+            for (const fileId of body.fileIds) {
+                const newKey = fileKeyMapping[fileId];
+                if (newKey) {
+                    // Insert into lead_files with updated r2_key
+                    await c.env.DB.prepare(`
+            INSERT INTO lead_files (id, site_id, lead_id, file_type, file_name, file_size, mime_type, r2_key, uploaded_at)
+            SELECT id, site_id, ?, file_type, file_name, file_size, mime_type, ?, uploaded_at
+            FROM staged_files
+            WHERE id = ? AND site_id = ?
+          `).bind(lead.id, newKey, fileId, siteId).run();
+                    // Delete from staged_files
+                    await c.env.DB.prepare(`DELETE FROM staged_files WHERE id = ? AND site_id = ?`)
+                        .bind(fileId, siteId).run();
+                }
+            }
+            console.log(`Associated ${body.fileIds.length} files with lead ${lead.id}`);
+        }
         return c.json({
             success: true,
             data: lead,
